backend/requirements_gemini.txt
# Google Gemini API for Summarization
google-generativeai>=0.3.0
